+++ database/migrations/2023_05_16_083058_add_ids_in_mention_users_table.php
@@ -16,21 +16,25 @@ return new class extends Migration
             $table->dropColumn('early_clock_in');
         });
 
-        Schema::table('employee_shifts', function (Blueprint $table) {
-            $table->string('early_clock_in')->nullable();
-        });
+        if (!Schema::hasColumn('employee_shifts', 'early_clock_in')) {
+            Schema::table('employee_shifts', function (Blueprint $table) {
+                $table->string('early_clock_in')->nullable();
+            });
+        }
 
-        Schema::table('mention_users', function (Blueprint $table) {
-            $table->integer('ticket_id')->unsigned()->nullable()->after('discussion_id');
-            $table->foreign('ticket_id')->references('id')->on('tickets')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
+        if (!Schema::hasColumn('mention_users', 'ticket_id')) {
+            Schema::table('mention_users', function (Blueprint $table) {
+                $table->integer('ticket_id')->unsigned()->nullable()->after('discussion_id');
+                $table->foreign('ticket_id')->references('id')->on('tickets')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
 
-            $table->integer('event_id')->unsigned()->nullable()->after('ticket_id');
-            $table->foreign('event_id')->references('id')->on('events')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-        });
+                $table->integer('event_id')->unsigned()->nullable()->after('ticket_id');
+                $table->foreign('event_id')->references('id')->on('events')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
